more on test and conversion

+++ exercises/tests/tests5.rs
@@ -31,7 +31,7 @@ unsafe fn modify_by_address(address: usize) {
     // code's behavior and the contract of this function. You may use the
     // comment of the test below as your format reference.
     unsafe {
-        let addr = address as *mut usize;
+        let addr = address as *mut u32;
         *addr = 0xAABBCCDD;
     }
 }
